feat: move deploy-command into a separate module
