+++ src/scripts/delete-global-commands.ts
@@ -1,6 +1,6 @@
 import dotenv from 'dotenv';
 import dotenvExpand from 'dotenv-expand';
-import { deployGlobalCommands } from '../commands/builder';
+import { deployGlobalCommands } from '../commands/deploy-command';
 
 const env = dotenv.config();
 dotenvExpand.expand(env);
